requirements.txt
# Web Framework
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-Login>=0.6.0
Flask-Bcrypt>=1.0.0
Werkzeug>=2.3.0

# Socket.IO
python-engineio>=4.5.0
python-socketio>=5.8.0

# Database
mysql-connector-python>=8.0.0

# AI/ML - PyTorch (GPU CUDA 12.4 版本)
# 安裝指令: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
# CPU 版本: pip install torch torchvision
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
